fixed some bugs, (logical indexing)
still have to fix order changes bug

+++ p2/stump_train.m
@@ -25,12 +25,12 @@ function [best_stump, ypred] = stump_train(Xtrain,Ytrain,D_t)
         
         for thres_idx=1:m-1
             threshold = (stump(thres_idx)+stump(thres_idx+1))/2;
-            pred_geq = (stump>=threshold);
+            pred_geq = double(stump>=threshold);
             pred_geq(pred_geq==0) = -1;
             err_geq = sum(D_t_corres(pred_geq~=stump_ytrain));
             assert(numel(err_geq)==1);
             
-            pred_leq = (stump<=threshold);
+            pred_leq = double(stump<=threshold);
             pred_leq(pred_leq==0) = -1;
             err_leq = sum(D_t_corres(pred_leq~=stump_ytrain));
             assert(numel(err_leq)==1);
@@ -51,7 +51,7 @@ function [best_stump, ypred] = stump_train(Xtrain,Ytrain,D_t)
                 best_thres = threshold;
                 best_direc = stump_candidate_dir;
                 best_err = stump_candidate_err;
-                ypred = stump_candidate_pred;
+                ypred = double(stump_candidate_pred);
             end         
         end
     end
